Update Ongoing Patient Medication updates the notifications
