+++ app/src/main/java/com/triadss/doctrack2/activity/patient/PatientHome.java
@@ -5,10 +5,15 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
+import androidx.work.Data;
+import androidx.work.OneTimeWorkRequest;
 import androidx.work.WorkManager;
 
+import android.content.Context;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.os.Bundle;
+import android.util.Log;
 import android.widget.Button;
 
 import com.google.firebase.auth.FirebaseAuth;
@@ -21,7 +26,17 @@ import com.triadss.doctrack2.activity.patient.fragments.PatientHomeFragment;
 import com.triadss.doctrack2.activity.patient.fragments.medications.PatientMedicationFragment;
 import com.triadss.doctrack2.activity.patient.fragments.records.PatientReportFragment;
 import com.triadss.doctrack2.activity.patient.fragments.records.RecordFragment;
+import com.triadss.doctrack2.config.constants.NotificationConstants;
+import com.triadss.doctrack2.config.constants.SessionConstants;
 import com.triadss.doctrack2.databinding.ActivityPatientHomeBinding;
+import com.triadss.doctrack2.dto.DateTimeDto;
+import com.triadss.doctrack2.dto.MedicationDto;
+import com.triadss.doctrack2.notification.NotificationMedicationScheduleWorker;
+import com.triadss.doctrack2.repoositories.MedicationRepository;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.TimeUnit;
 
 public class PatientHome extends AppCompatActivity {
 
@@ -30,6 +45,8 @@ public class PatientHome extends AppCompatActivity {
     Button button;
 
     FirebaseUser user;
+    private String loggedInUserId;
+    MedicationRepository medicationRepository = new MedicationRepository();
 
     ActivityPatientHomeBinding binding;
 
@@ -38,6 +55,10 @@ public class PatientHome extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_patient_home);
         replaceFragment(new PatientHomeFragment());
+
+        SharedPreferences sharedPref = getSharedPreferences(SessionConstants.SessionPreferenceKey,
+                Context.MODE_PRIVATE);
+        loggedInUserId = sharedPref.getString(SessionConstants.LoggedInUid, "");
 
         auth = FirebaseAuth.getInstance();
         user = auth.getCurrentUser();
@@ -103,5 +124,48 @@ public class PatientHome extends AppCompatActivity {
             fragmentTransaction.addToBackStack("toHome");
         }
         fragmentTransaction.commit();
+    }
+
+    public void setupNotifications() {
+        medicationRepository.getUncompletedMedications(loggedInUserId, new MedicationRepository.MedicationFetchCallback() {
+            @Override
+            public void onSuccess(List<MedicationDto> medications) {
+                List<OneTimeWorkRequest> requests = new ArrayList<OneTimeWorkRequest>();
+                for(MedicationDto dto : medications) {
+                    Log.e("TEST", "Running Medication Work for " + loggedInUserId +
+                            " for " + dto.getMedicine() + " at " + DateTimeDto.ToDateTimeDto(dto.getTimestamp()).ToString());
+
+                    long distance = DateTimeDto.GetTimestampDiffInSeconds(dto.getTimestamp());
+
+                    OneTimeWorkRequest notifWorkRequest = new OneTimeWorkRequest.Builder(
+                            NotificationMedicationScheduleWorker.class)
+                            .setInputData(new Data.Builder()
+                                    .putString(NotificationConstants.RECEIVER_ID, loggedInUserId)
+                                    .putString(NotificationConstants.TITLE_ID, String.format("Please Take your medication %s", dto.getMedicine()))
+                                    .putString(NotificationConstants.CONTENT_ID, String.format("Medication should be taken at %s", DateTimeDto.ToDateTimeDto(dto.getTimestamp()).ToString()))
+                                    .build())
+                            .addTag(NotificationConstants.MEDICATION_NOTIFICATION_TAG)
+                            .setInitialDelay(distance, TimeUnit.SECONDS)
+                            .build();
+
+                    requests.add(notifWorkRequest);
+                }
+                WorkManager.getInstance(PatientHome.this)
+                        .cancelAllWorkByTag(NotificationConstants.MEDICATION_NOTIFICATION_TAG);
+
+                Log.e("TEST", "Medication notif count " + String.valueOf(requests.size()));
+
+
+                if(!requests.isEmpty()) {
+                    WorkManager.getInstance(PatientHome.this)
+                            .enqueue(requests);
+                }
+            }
+
+            @Override
+            public void onError(String errorMessage) {
+
+            }
+        });
     }
 }
